+++ Scripts/Player/PlayerInteractionTool.cs
@@ -837,7 +837,7 @@ namespace VoxelEngine.Player
                 Vector3 origin = ray.origin + ray.direction * 1.2f;
                 Vector3 throwVel = ray.direction * weapon.throwForce + up * 3.2f; // forward + upward arc
                 VoxelEngine.Combat.BombProjectile.Spawn(origin, throwVel, gameObject, weapon.explosionMaterial,
-                    weapon.explosionRadius, weapon.explosionDamage, weapon.fuseTime, weapon.explosionMaterial);
+                    weapon.explosionRadius, weapon.explosionDamage, weapon.fuseTime, weapon.voxelDamageRadius, weapon.explosionMaterial);
                 GetComponent<VoxelEngine.Player.HeldToolView>()?.DoRecoil();
                 inventory.container.Remove(weapon, 1);   // consumable: spend one grenade
                 inventory.container.RaiseChanged();